Refactor test cases into two files
