+++ src/test/resources/sniperPrinter/tryWithResource/PrintOnce.java
@@ -1,0 +1,12 @@
+package sniperPrinter.tryWithResource;
+
+import java.util.zip.ZipFile;
+import java.io.BufferedWriter;
+import java.nio.file.Files.newBufferedWriter;
+
+public class PrintOnce {
+    public void resourcePrinting() {
+        try (ZipFile zf = new ZipFile(zipFileName);
+             BufferedWriter writer = newBufferedWriter(outputFilePath, charset)) { }
+    }
+}
