+++ src/test/resources/sniperPrinter/tryWithResource/RetainSemiColon.java
@@ -1,13 +1,11 @@
-package sniperPrinter;
+package sniperPrinter.tryWithResource;
 
 import java.util.zip.ZipFile;
 import java.io.BufferedWriter;
 import java.nio.file.Files.newBufferedWriter;
 
-public class TryWithResource {
+public class RetainSemiColon {
     public void resourcePrinting() {
-        try (ZipFile zf = new ZipFile(zipFileName);
-             BufferedWriter writer = newBufferedWriter(outputFilePath, charset)) { }
         try (ZipFile zf = new ZipFile(zipFileName);
              BufferedWriter writer = newBufferedWriter(outputFilePath, charset);) { }
     }
